restructure folder & move files
